SPRs 311, 313. Forward Analysis
SPRs 313 - Eliminate from the registry algorithms not loaded
SPRs 311 - Eliminate occurrence of “Python Console” item in the Tools menu

Forward Analysis
- Began with scatters
- Extension to IDEs (Windows & mac incomplete).
- build scripts update
- Chart tests
- 2015 compilation
- qt5 port: code and tools (incomplete)
- project templates for qt5
- Updated notes
- Removed generated files from version control.
- Added VS projects

General
- Added updated SUM
- Moved SUM.doc
- opengl registry scripts for windows 10 (again, do not work)
- Correction of externals bug from v3
- Script to build Command Line Tools in CentOS 7 - 64 bit

+++ new-gui/Common/GUI/ProcessesTable.cpp
@@ -513,7 +513,7 @@ const QString& CProcessesTable::ProcessErrorMessage( QProcess::ProcessError erro
 
 	static_assert( ( QProcess::ProcessError::UnknownError + 1 ) == msgs_size, "ProcessError enumerated values differ in size from their respective messages array." );
 
-	assert__( error < msgs_size );
+	assert__( error < (int)msgs_size );
 
 	return msgs[ error ];
 }
